commit 3/4/2021

+++ web-film/src/main/java/com/movie/web/api/CommentAPI.java
@@ -1,25 +1,42 @@
 package com.movie.web.api;
 
+import com.movie.core.constant.WebConstant;
 import com.movie.core.dto.CommentDTO;
+import com.movie.core.dto.UserDTO;
 import com.movie.core.service.ICommentService;
 import com.movie.web.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController(value = "CommentAPIOfWeb")
 public class CommentAPI {
     @Autowired
     private ICommentService commentService;
+
     @PostMapping("/api/film/comment")
-    public void saveComment(@RequestBody CommentDTO commentDTO) {
-        commentService.save(commentDTO);
+    public Long saveComment(@RequestBody CommentDTO commentDTO) throws Exception {
+        UserDTO userDTO = new UserDTO();
+        userDTO.setId(SecurityUtils.getUserPrincipal().getId());
+        commentDTO.setUser(userDTO);
+        return commentService.save(commentDTO);
     }
 
+    // chức năng delete chỉ sử dụng được với 1 id, nhiều id chưa sửa :))
     @DeleteMapping("/api/film/comment")
     public void deleteComment(@RequestBody Long[] ids) {
-        commentService.deleteComment(ids);
+        if (SecurityUtils.getEmployeeAuthorities().contains(WebConstant.ROLE_ADMIN)) {
+            commentService.deleteComment(ids);
+        } else if (SecurityUtils.getEmployeeAuthorities().contains(WebConstant.ROLE_POSTER)) {
+            CommentDTO commentDTO = commentService.findOneById(ids[0]);
+            if (commentDTO.getFilm().getEmployee().getId() == SecurityUtils.getPrincipal().getId()) {
+                commentService.deleteComment(ids);
+            }
+        } else if (SecurityUtils.getUserAuthorities().contains(WebConstant.ROLE_USER)) {
+            CommentDTO commentDTO = commentService.findOneById(ids[0]);
+            if (commentDTO.getUser().getId() == SecurityUtils.getUserPrincipal().getId()) {
+                commentService.deleteComment(ids);
+            }
+        }
     }
 
     @GetMapping("/api/admin/comment/content")
